+++ wcomponents-examples/src/main/java/com/github/bordertech/wcomponents/examples/WAudioExample.java
@@ -4,6 +4,7 @@ import com.github.bordertech.wcomponents.Action;
 import com.github.bordertech.wcomponents.ActionEvent;
 import com.github.bordertech.wcomponents.Audio;
 import com.github.bordertech.wcomponents.AudioResource;
+import com.github.bordertech.wcomponents.HeadingLevel;
 import com.github.bordertech.wcomponents.Margin;
 import com.github.bordertech.wcomponents.Request;
 import com.github.bordertech.wcomponents.Size;
@@ -13,11 +14,13 @@ import com.github.bordertech.wcomponents.WButton;
 import com.github.bordertech.wcomponents.WCheckBox;
 import com.github.bordertech.wcomponents.WContainer;
 import com.github.bordertech.wcomponents.WFieldLayout;
+import com.github.bordertech.wcomponents.WHeading;
 import com.github.bordertech.wcomponents.WLabel;
-import com.github.bordertech.wcomponents.subordinate.Disable;
-import com.github.bordertech.wcomponents.subordinate.Enable;
+import com.github.bordertech.wcomponents.WTextField;
 import com.github.bordertech.wcomponents.subordinate.Equal;
+import com.github.bordertech.wcomponents.subordinate.Hide;
 import com.github.bordertech.wcomponents.subordinate.Rule;
+import com.github.bordertech.wcomponents.subordinate.Show;
 import com.github.bordertech.wcomponents.subordinate.WSubordinateControl;
 
 /**
@@ -44,14 +47,21 @@ public class WAudioExample extends WContainer {
 	private final WCheckBox cbLoop = new WCheckBox();
 
 	/**
-	 * Used to set the controls option on audio to PLAY_PAUSE.
+	 * Used to set the muted status.
 	 */
-	private final WCheckBox cbControls = new WCheckBox();
+	private final WCheckBox cbMute = new WCheckBox();
 
 	/**
-	 * Used to set the disabled option on audio.
+	 * Show/hide controller.
 	 */
-	private final WCheckBox cbDisable = new WCheckBox();
+	private final WCheckBox cbHide = new WCheckBox();
+
+	private final WCheckBox cbRenderControls = new WCheckBox(true);
+
+	/**
+	 * Used to set mediagroup attribute.
+	 */
+	private final WTextField tfMediaGroup = new WTextField();
 
 	/**
 	 * A button used to apply new settings to the audio component.
@@ -73,7 +83,6 @@ public class WAudioExample extends WContainer {
 				};
 
 		audio.setAudio(audioClips);
-		audio.setAltText("Example WAudio content");
 		btnApply.setAction(new Action() {
 			@Override
 			public void execute(final ActionEvent event) {
@@ -94,24 +103,32 @@ public class WAudioExample extends WContainer {
 		add(layout);
 		layout.addField("Autoplay", cbAutoPlay);
 		layout.addField("Loop", cbLoop);
-		layout.addField("Disable", cbDisable);
-		layout.addField("Show only play/pause", cbControls);
+		layout.addField("Mute on load", cbMute);
+		layout.addField("Hide audio", cbHide);
+		layout.addField("render default controls", cbRenderControls);
+		layout.addField("Media Group", tfMediaGroup);
 		layout.addField((WLabel) null, btnApply);
 
 		// enable disable option only when control PLAY_PAUSE is used.
 		WSubordinateControl control = new WSubordinateControl();
 		add(control);
 		Rule rule = new Rule();
-		rule.setCondition(new Equal(cbControls, Boolean.TRUE.toString()));
-		rule.addActionOnTrue(new Enable(cbDisable));
-		rule.addActionOnFalse(new Disable(cbDisable));
+		rule.setCondition(new Equal(cbHide, Boolean.TRUE.toString()));
 		control.addRule(rule);
+		rule.addActionOnTrue(new Hide(audio));
+		rule.addActionOnFalse(new Show(audio));
+
 
 		// allow config to change without reloading the whole page.
 		add(new WAjaxControl(btnApply, audio));
 
 		// add the audio to the UI
 		add(audio);
+
+		add(new WHeading(HeadingLevel.H2, "Audio with single source"));
+		add(new WAudio("/audio/ogg.ogg"));
+
+
 	}
 
 	/**
@@ -133,7 +150,9 @@ public class WAudioExample extends WContainer {
 	private void setupAudio() {
 		audio.setAutoplay(cbAutoPlay.isSelected());
 		audio.setLoop(!cbLoop.isDisabled() && cbLoop.isSelected());
-		audio.setControls(cbControls.isSelected() ? WAudio.Controls.PLAY_PAUSE : WAudio.Controls.NATIVE);
-		audio.setDisabled(cbControls.isSelected() && cbDisable.isSelected());
+		audio.setMuted(cbMute.isSelected());
+		audio.setRenderControls(cbRenderControls.isSelected());
+		String mediaGroup = tfMediaGroup.getText();
+		audio.setMediaGroup("".equals(mediaGroup) ? null : mediaGroup);
 	}
 }